Updates and Bug Fixes
Adjusted camera parent naming. Since the program does not differentiate between any of the camera types when rendering they have been renamed to Camera 1, Camera 2 etc.
Fixed a bug with NLA Action Strip frame counts. The code used the original action start and end frames, not the start and end frames as set in the NLA Editor as it should have.